refactor(tls-harness): separate benchmark abstractions
