+++ bindings/rust/standard/tls-harness/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "bench"
+name = "tls-harness"
 version = "0.1.0"
 edition = "2021"
 publish = false
