+++ bindings/rust/standard/tls-harness/benches/handshake.rs
@@ -1,15 +1,13 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-use bench::{
-    harness::{TlsBenchConfig, TlsInfo},
-    CipherSuite, CryptoConfig, HandshakeType, KXGroup, Mode, OpenSslConnection, RustlsConnection,
-    S2NConnection, SigType, TlsConnPair, TlsConnection,
-};
 use criterion::{
     criterion_group, criterion_main, measurement::WallTime, BatchSize, BenchmarkGroup, Criterion,
 };
 use strum::IntoEnumIterator;
+use tls_harness::bench_config::*;
+use tls_harness::cohort::{OpenSslConnection, RustlsConnection, S2NConnection};
+use tls_harness::{harness::TlsInfo, Mode, SigType, TlsConnPair, TlsConnection};
 
 fn bench_handshake_for_library<T>(
     bench_group: &mut BenchmarkGroup<WallTime>,
