+++ bindings/rust/standard/tls-harness/certs/ecdsa256
