+++ bindings/rust/standard/tls-harness/src/bench_config/mod.rs
@@ -1,0 +1,133 @@
+//! This module holds all of the "benchmark specific" configuration logic that is
+//! used in the benchmark suites.
+//!
+//! Anything that is only relevant to the benchmarks should live in this module.
+//! Similarly, anything that might be used outside of the benchmarks should _not_
+//! live in this module.
+
+use std::{error::Error, fmt::Debug};
+use strum::EnumIter;
+
+mod setup;
+
+use crate::{Mode, SigType, TlsConnPair, TlsConnection};
+
+/// While ServerAuth and Resumption are not mutually exclusive, they are treated
+/// as such for the purpose of benchmarking.
+#[derive(Clone, Copy, Default, EnumIter, Eq, PartialEq)]
+pub enum HandshakeType {
+    #[default]
+    ServerAuth,
+    MutualAuth,
+    Resumption,
+}
+
+impl Debug for HandshakeType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            HandshakeType::ServerAuth => write!(f, "server-auth"),
+            HandshakeType::MutualAuth => write!(f, "mTLS"),
+            HandshakeType::Resumption => write!(f, "resumption"),
+        }
+    }
+}
+
+// these parameters were the only ones readily usable for all three libaries:
+// s2n-tls, rustls, and openssl
+#[allow(non_camel_case_types)]
+#[derive(Clone, Copy, Debug, Default, EnumIter, Eq, PartialEq)]
+pub enum CipherSuite {
+    #[default]
+    TLS_AES_128_GCM_SHA256,
+    TLS_AES_256_GCM_SHA384,
+}
+
+#[derive(Clone, Copy, Default, EnumIter)]
+pub enum KXGroup {
+    Secp256R1,
+    #[default]
+    X25519,
+}
+
+impl Debug for KXGroup {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::Secp256R1 => write!(f, "secp256r1"),
+            Self::X25519 => write!(f, "x25519"),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default)]
+pub struct CryptoConfig {
+    pub cipher_suite: CipherSuite,
+    pub kx_group: KXGroup,
+    pub sig_type: SigType,
+}
+
+impl CryptoConfig {
+    pub fn new(cipher_suite: CipherSuite, kx_group: KXGroup, sig_type: SigType) -> Self {
+        Self {
+            cipher_suite,
+            kx_group,
+            sig_type,
+        }
+    }
+}
+
+/// The TlsBenchConfig trait allows us to map benchmarking parameters to
+/// a configuration object
+pub trait TlsBenchConfig: Sized {
+    fn make_config(
+        mode: Mode,
+        crypto_config: CryptoConfig,
+        handshake_type: HandshakeType,
+    ) -> Result<Self, Box<dyn Error>>;
+}
+
+impl<C, S> TlsConnPair<C, S>
+where
+    C: TlsConnection,
+    S: TlsConnection,
+    C::Config: TlsBenchConfig,
+    S::Config: TlsBenchConfig,
+{
+    /// Initialize buffers, configs, and connections (pre-handshake)
+    pub fn new_bench_pair(
+        crypto_config: CryptoConfig,
+        handshake_type: HandshakeType,
+    ) -> Result<Self, Box<dyn Error>> {
+        // do an initial handshake to generate the session ticket
+        if handshake_type == HandshakeType::Resumption {
+            let server_config =
+                S::Config::make_config(Mode::Server, crypto_config, handshake_type)?;
+            let client_config =
+                C::Config::make_config(Mode::Client, crypto_config, handshake_type)?;
+
+            // handshake the client and server connections. This will result in
+            // session ticket getting stored in client_config
+            let mut pair = TlsConnPair::<C, S>::from_configs(&client_config, &server_config);
+            pair.handshake()?;
+            // NewSessionTicket messages are part of the application data and sent
+            // after the handshake is complete, so we must trigger an additional
+            // "read" on the client connection to ensure that the session ticket
+            // gets received and stored in the config
+            pair.round_trip_transfer(&mut [0]).unwrap();
+
+            // new_from_config is called interally by the TlsConnPair::new
+            // method and will check if a session ticket is available and set it
+            // on the connection. This results in the session ticket in
+            // client_config (from the previous handshake) getting set on the
+            // client connection.
+            return Ok(TlsConnPair::<C, S>::from_configs(
+                &client_config,
+                &server_config,
+            ));
+        }
+
+        Ok(TlsConnPair::<C, S>::from_configs(
+            &C::Config::make_config(Mode::Client, crypto_config, handshake_type).unwrap(),
+            &S::Config::make_config(Mode::Server, crypto_config, handshake_type).unwrap(),
+        ))
+    }
+}
