+++ bindings/rust/standard/tls-harness/src/harness/mod.rs
@@ -75,79 +75,6 @@ pub enum Mode {
     Server,
 }
 
-/// While ServerAuth and Resumption are not mutually exclusive, they are treated
-/// as such for the purpose of benchmarking.
-#[derive(Clone, Copy, Default, EnumIter, Eq, PartialEq)]
-pub enum HandshakeType {
-    #[default]
-    ServerAuth,
-    MutualAuth,
-    Resumption,
-}
-
-impl Debug for HandshakeType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            HandshakeType::ServerAuth => write!(f, "server-auth"),
-            HandshakeType::MutualAuth => write!(f, "mTLS"),
-            HandshakeType::Resumption => write!(f, "resumption"),
-        }
-    }
-}
-
-// these parameters were the only ones readily usable for all three libaries:
-// s2n-tls, rustls, and openssl
-#[allow(non_camel_case_types)]
-#[derive(Clone, Copy, Debug, Default, EnumIter, Eq, PartialEq)]
-pub enum CipherSuite {
-    #[default]
-    AES_128_GCM_SHA256,
-    AES_256_GCM_SHA384,
-}
-
-#[derive(Clone, Copy, Default, EnumIter)]
-pub enum KXGroup {
-    Secp256R1,
-    #[default]
-    X25519,
-}
-
-impl Debug for KXGroup {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Secp256R1 => write!(f, "secp256r1"),
-            Self::X25519 => write!(f, "x25519"),
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, Default)]
-pub struct CryptoConfig {
-    pub cipher_suite: CipherSuite,
-    pub kx_group: KXGroup,
-    pub sig_type: SigType,
-}
-
-impl CryptoConfig {
-    pub fn new(cipher_suite: CipherSuite, kx_group: KXGroup, sig_type: SigType) -> Self {
-        Self {
-            cipher_suite,
-            kx_group,
-            sig_type,
-        }
-    }
-}
-
-/// The TlsBenchConfig trait allows us to map benchmarking parameters to
-/// a configuration object
-pub trait TlsBenchConfig: Sized {
-    fn make_config(
-        mode: Mode,
-        crypto_config: CryptoConfig,
-        handshake_type: HandshakeType,
-    ) -> Result<Self, Box<dyn Error>>;
-}
-
 /// The TlsConnection object can be created from a corresponding config type.
 pub trait TlsConnection: Sized {
     /// Library-specific config struct
@@ -192,7 +119,11 @@ pub trait TlsConnection: Sized {
 
 pub trait TlsInfo: Sized {
     fn name() -> String;
-    fn get_negotiated_cipher_suite(&self) -> CipherSuite;
+
+    /// Return the IANA Description of the negotiated cipher suite.
+    ///
+    /// e.g. `TLS_AES_256_GCM_SHA384`
+    fn get_negotiated_cipher_suite(&self) -> String;
 
     fn negotiated_tls13(&self) -> bool;
 
@@ -208,65 +139,6 @@ pub struct TlsConnPair<C, S> {
     pub client: C,
     pub server: S,
     pub io: TestPairIO,
-}
-
-impl<C, S> Default for TlsConnPair<C, S>
-where
-    C: TlsConnection,
-    S: TlsConnection,
-    C::Config: TlsBenchConfig,
-    S::Config: TlsBenchConfig,
-{
-    fn default() -> Self {
-        Self::new_bench_pair(CryptoConfig::default(), HandshakeType::default()).unwrap()
-    }
-}
-
-impl<C, S> TlsConnPair<C, S>
-where
-    C: TlsConnection,
-    S: TlsConnection,
-    C::Config: TlsBenchConfig,
-    S::Config: TlsBenchConfig,
-{
-    /// Initialize buffers, configs, and connections (pre-handshake)
-    pub fn new_bench_pair(
-        crypto_config: CryptoConfig,
-        handshake_type: HandshakeType,
-    ) -> Result<Self, Box<dyn Error>> {
-        // do an initial handshake to generate the session ticket
-        if handshake_type == HandshakeType::Resumption {
-            let server_config =
-                S::Config::make_config(Mode::Server, crypto_config, handshake_type)?;
-            let client_config =
-                C::Config::make_config(Mode::Client, crypto_config, handshake_type)?;
-
-            // handshake the client and server connections. This will result in
-            // session ticket getting stored in client_config
-            let mut pair = TlsConnPair::<C, S>::from_configs(&client_config, &server_config);
-            pair.handshake()?;
-            // NewSessionTicket messages are part of the application data and sent
-            // after the handshake is complete, so we must trigger an additional
-            // "read" on the client connection to ensure that the session ticket
-            // gets received and stored in the config
-            pair.round_trip_transfer(&mut [0]).unwrap();
-
-            // new_from_config is called interally by the TlsConnPair::new
-            // method and will check if a session ticket is available and set it
-            // on the connection. This results in the session ticket in
-            // client_config (from the previous handshake) getting set on the
-            // client connection.
-            return Ok(TlsConnPair::<C, S>::from_configs(
-                &client_config,
-                &server_config,
-            ));
-        }
-
-        Ok(TlsConnPair::<C, S>::from_configs(
-            &C::Config::make_config(Mode::Client, crypto_config, handshake_type).unwrap(),
-            &S::Config::make_config(Mode::Server, crypto_config, handshake_type).unwrap(),
-        ))
-    }
 }
 
 impl<C, S> TlsConnPair<C, S>
@@ -358,7 +230,7 @@ where
     C: TlsInfo,
     S: TlsInfo,
 {
-    pub fn get_negotiated_cipher_suite(&self) -> CipherSuite {
+    pub fn get_negotiated_cipher_suite(&self) -> String {
         assert!(
             self.client.get_negotiated_cipher_suite() == self.server.get_negotiated_cipher_suite()
         );
@@ -373,7 +245,10 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{OpenSslConnection, RustlsConnection, S2NConnection, TlsConnPair};
+    use crate::{
+        bench_config::{CryptoConfig, HandshakeType, TlsBenchConfig},
+        cohort::{OpenSslConnection, RustlsConnection, S2NConnection},
+    };
     use std::{io::ErrorKind, path::Path};
     use strum::IntoEnumIterator;
