+++ bindings/rust/standard/tls-harness/src/lib.rs
@@ -1,0 +1,13 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod bench_config;
+pub mod cohort;
+pub mod harness;
+#[cfg(test)]
+pub mod test_utilities;
+
+pub use crate::harness::{get_cert_path, Mode, PemType, SigType, TlsConnPair, TlsConnection};
+
+// controls profiler frequency for flamegraph generation in benchmarks
+pub const PROFILER_FREQUENCY: i32 = 100;
